requirements.txt
beautifulsoup4>=4.10.0
cloudscraper==1.2.69
Flask>=2.0.1
loguru>=0.7.0
discord.py>=2.3.2
python-dotenv>=0.17.1
requests>=2.27.1
requests-toolbelt>=0.9.1
typing_extensions>=4.1.1
